summarizer_sentiment_service/requirements.txt
# ctransformers
# sentence-transformers
# langchain
# langchain-community
python-dotenv
google-generativeai
fastapi
uvicorn

